+++ docs/Changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## Beta Updates
+* Reduce Menu loading logic
+* Load js injectors after showing webview
+* Add notifications for messages
+
+## v1.3
 * Create toggle for notifications only from primary account
 * Micro string optimizations
 * Add profile icons to notifications